requirements.txt
Brotli==1.0.9
click==8.1.3
colorama==0.4.4
dash==2.5.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.1.2
Flask-Compress==1.12
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
pandas==1.4.2
plotly==5.8.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
tenacity==8.0.1
Werkzeug==2.1.2
zipp==3.8.0
boto3==1.24.13
python-dotenv==0.20.0
scipy==1.8.1
dash-bootstrap-component==1.1.0